Add files via upload
input: expression ended with =
output: value with test information

This is clumsy attemp to write a program which can caluculate polynomial fuction. This version could only recognize fuction with no parenthese. 
Despite of complex code and terrible programming habit, I am so proud of it. And next time, I will be better.