keep going!
cant get an idea why i cant find a job. As a mechanical major graduate, i cant find either a mechanical job or a CS base job. ever had i doubted self-value and the meaning of graduate life. I confirm the great effort and tough time during 2 years in Northeastern. I believe that i dont waste time. So why I cant get a equal-paid job and a normal girlfriend?
Maybe I was in the wrong way. Maybe what i learned is to make me a qualified tool and i cant find a place to fit in. 
Fine, complains and the past memory are useless pains for a man. I drops all the recreation, procastination, laziness and illusion of love for a second change to reach my top!